+++ models/src/main/java/ar/edu/itba/paw/model/Category.java
@@ -1,19 +1,19 @@
 package ar.edu.itba.paw.model;
 
 public class Category {
-    private final int categoryId;
+    private final long categoryId;
     private final Restaurant restaurant;
     private final String name;
     private final int order;
 
-    public Category(int categoryId, Restaurant restaurant, String name, int order) {
+    public Category(long categoryId, Restaurant restaurant, String name, int order) {
         this.categoryId = categoryId;
         this.restaurant = restaurant;
         this.name = name;
         this.order = order;
     }
 
-    public int getCategoryId() {
+    public long getCategoryId() {
         return categoryId;
     }
